Cleaned up installation process and files
The Install.sh script handles a wide variety of installation
cases, including creating a linking .emacs file.

The Update.sh script is for the case where the active
.emacs.d directory is different from the repository
and we want to update extra files like Cask.

The Cask file has been improved to eliminate a few
obsolete packages, and the home-dot-emacs.el now
let's one choose a package approach.

The README describes this briefly but could be improved.