Separated sections in dot-emacs to their own files; many cleanups
* Frames, keybindings, translations, hooks, etc. in separate files
* Tools and languages separated into their own components
* Changed how the list of mods is referenced
* Cleaned up dot-emacs to respect preferences and be easier to navigate
* Fixed absolute target bug in Install.sh
* Added user and system data to data/ where it makes sense
* Renamed files to eliminate crg- prefix
* Small documentation fixes

The absolute target bug occured when putting the target name
in the home script. A relative name would be faulty as the
relative position is not the same; converted to an absolute
target path for the sed step.

The distinction between mods and components is somewhat fluid,
but what I am holding to is that mods represent structural
changes or code that offers new/improved functionality whereas
components simply load and configure a cluster of related
packages and modes.