Fixed paredit shadowing and added completion tools
The structure of the paredit keymap was causing the original
keybindings to be maintained in the map despite resetting
of paredit-commands. I fixed this with a new function to
reset the commands while respecting minor-mode-map-alist,
allowing the reset to happen after the mode is defined.

Defined the ivy yank action, which I used but had forgotten.
Added counsel-find-file-read-only to be placed in the
global keymap. Still investigating how generally those
operations can be controlled with counsel/ivy (for instance,
to filter out backup files).