Improved upgraded Install.sh, fixed README, added tutorial.org
Install.sh now handles temporary files and command substitution
more portably and cleanly. This should help avoid problems
on windows, for instance. The various ways of using this script
are now better documented in the README.org file.

The tutorial.org file, reviewed with the review tool during
installation, is now more complete and useful. This is probably
a work in progress, however.