Fixed review customization and package installation
Loading the custom file does not allow effective customization
unless the corresponding packages are loaded to give the
customize values the proper settings and values. So, the
dot-emacs was re-arranged slightly to separate the startup
actions for cold loading, and the review loads the local
configuration with settings that keep everything in whatever
directory is being considered (possibly not the user's emacs
directory). This enables meaningful customizations, and it has
the beneficial effect of making the review tool match the
eventual environment to be used. Everything looks better,
except the initial frame is kept small.

The package installation required some work to initialize properly
and keep everything local to the config directory being used
(which might not be the user's emacs dir). Added specialized
fixes, including the auctex provide change. This works well,
though it takes a long while. Should output a message to
that effect and should have the verbose flag to the Install
script filter out just my messages (ignoring compilation etc)
unless set.  (| $filter, where filter is grep ... or cat).

Small fixes to tutorial to add value and to emacs-custon
to eliminate some cruft.