Simplified the options used by Install.sh.
Made the full install the default, with a variety
of --no-* and --force-* options to flexibly handle
special cases. Improved handling of the no packages
installed possibilities.

Next task is to filter the output of a package-type
package install based on --verbose setting. But
for the time being, it should be ok.