In cask case, let review find packages, Install do specialty fixes
Previous version had forgotten different package path
for review loading when cask or cask-homebrew is the
package type. Fixed this by specifying it to the review
entry point and initialized the proper package manager.

Also, cask install did not do specialty fixes, but
used that facility in Extras/packages.el to do that.
Used the cask emacs command to keep everything consistent.